refactor(web): split remote agent gateway handlers [OPE-362]
